change int to Integer

+++ backend/back-office-api/src/main/java/org/academiadecodigo/bootcamp65/model/User.java
@@ -3,27 +3,27 @@ package org.academiadecodigo.bootcamp65.model;
 import java.util.List;
 
 public class User {
-    private int id;
+    private Integer id;
     private String img;
     private String name;
-    private int age;
+    private Integer age;
     private GenderType gender;
 
     private int minAge;
     private int maxAge;
 
-    private int lat;
-    private int lon;
+    private double lat;
+    private double lon;
 
-    private int rating;
-    private int time;
-    private int duration;
+    private Integer rating;
+    private Integer time;
+    private Integer duration;
 
-    public int getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -43,11 +43,11 @@ public class User {
         this.name = name;
     }
 
-    public int getAge() {
+    public Integer getAge() {
         return age;
     }
 
-    public void setAge(int age) {
+    public void setAge(Integer age) {
         this.age = age;
     }
 
@@ -75,43 +75,43 @@ public class User {
         this.maxAge = maxAge;
     }
 
-    public int getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(int lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public int getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(int lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
 
-    public int getRating() {
+    public Integer getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(Integer rating) {
         this.rating = rating;
     }
 
-    public int getTime() {
+    public Integer getTime() {
         return time;
     }
 
-    public void setTime(int time) {
+    public void setTime(Integer time) {
         this.time = time;
     }
 
-    public int getDuration() {
+    public Integer getDuration() {
         return duration;
     }
 
-    public void setDuration(int duration) {
+    public void setDuration(Integer duration) {
         this.duration = duration;
     }
 }
